# cmd/server/openapi.yaml
openapi: 3.1.0
info:
  title: Echo API
  version: 1.0.0
  description: |
    The Echo feedback platform API.

    All request bodies are JSON; all responses are JSON. Authentication uses a
    Bearer JWT issued by `POST /v1/login` and carried in the `Authorization`
    header as `Bearer <access_token>`. Routes marked with a lock icon require
    this token; public routes do not.

    The dev server listens on `http://localhost:1323`.
servers:
  - url: http://localhost:1323
    description: Local development server
...
paths:
...
  /v1/register:
    post:
      tags: [Auth]
      summary: Register a new employee
      description: |
        Creates a new employee and returns the created record. The `password`
        is hashed server-side with bcrypt and never returned. A fresh account
        has `is_mail_verified: false` and `role: "org_admin"` when registered
        without an invitation. On success the server sends a verification email
        (logged to stdout when no mailer is configured).

        Authorization by invitation token:
          * With a non-empty `invite_token`: the token is verified and, on
            success, its organization name overrides whatever the client
            supplied, and the role is `"user"`. An invalid/expired token is
            rejected with `400 "invalid or expired invitation token"`.
          * Without a token: the client-supplied `organization_name` is
            required and honored, and the role is `"org_admin"` (the first
            admin). This is only allowed for an organization that does not
            exist yet: registering without a token into an organization that
            already has members returns `409 "organization already exists"`.
            Joining an existing organization requires an invitation token
            issued by one of its admins.
      operationId: registerEmployee
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterEmployeeRequest"
            examples:
              firstAdmin:
                summary: First admin (no invitation)
                value:
                  name: Alice
                  organization_name: Acme
                  manager_id: null
                  title: Senior Engineer
                  email: alice@example.com
                  password: supersecret
                  invite_token: ""
      responses:
        "201":
          description: The created employee. The `password` field is omitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: An employee with that email already exists, or the
            organization already has members and no invitation token was
            supplied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                emailTaken:
                  value: { message: "email already taken" }
                organizationTaken:
                  value: { message: "organization already exists" }
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/login:
    post:
      tags: [Auth]
      summary: Log in
      description: |
        Authenticates an employee by email and password and returns a signed
        JWT on success. Invalid email or wrong password both return
        `401 "invalid email or password"` to avoid leaking which was wrong.
        If the credentials are valid but the email is not verified, returns
        `403 "email not verified"`.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalid:
                  value: { message: "invalid email or password" }
        "403":
          description: Email not verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                unverified:
                  value: { message: "email not verified" }
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/verify-email:
    get:
      tags: [Auth]
      summary: Verify an email address
      description: |
        Validates the email-verification token supplied in the `token` query
        parameter and, on success, marks the matching employee's email as
        verified. Public and idempotent: verifying an already-verified email
        succeeds again with the same `200` response.
      operationId: verifyEmail
      security: []
      parameters:
        - $ref: "#/components/parameters/VerificationToken"
      responses:
        "200":
          description: Email verified.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required: [message]
              examples:
                ok:
                  value: { message: "email verified" }
        "400":
          description: Missing, malformed, expired, or wrong-purpose token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalid:
                  value: { message: "invalid or expired verification token" }
  /v1/me:
    get:
      tags: [Employees]
      summary: Get the current employee's profile
      description: |
        Returns the profile of the authenticated employee. The employee is
        looked up by the token's subject (the employee ID).
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        "200":
          description: The authenticated employee. The `password` field is omitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: The token's subject no longer matches a stored employee.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                notFound:
                  value: { message: "employee not found" }
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/employees:
    get:
      tags: [Employees]
      summary: List employees in the caller's organization
      description: |
        Returns one page of employees in the authenticated caller's
        organization, ordered by name ascending. An employee can only see
        their own organization's members. The response omits the `password`
        field. The caller is included in the list; the frontend should filter
        the caller out when populating a reviewee picker (the backend enforces
        no self-review at `POST /v1/feedbacks`).

        Pagination is cursor-based: pass the `next_cursor` value from one
        response as the `cursor` query parameter on the next request. When
        `next_cursor` is `null` there are no more pages. An unknown cursor
        (one that does not refer to an existing employee) returns `400`.
      operationId: listEmployees
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListCursor"
      responses:
        "200":
          description: One page of employees plus the next-page cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
        "400":
          description: Unknown cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                unknownCursor:
                  value: { message: "unknown cursor" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitation:
    post:
      tags: [Invitations]
      summary: Create an invitation token
      description: |
        Issues a signed invitation token that lets the bearer register as a
        member of the inviter's organization. The caller must have
        `role: "org_admin"`; any other role returns `403`. The organization is
        taken from the caller's JWT (not the body), so an admin cannot mint
        join-tokens for an organization they do not administer. The returned
        `token` is then passed as `invite_token` in `POST /v1/register` to
        create the invitee's account with `role: "user"`.
      operationId: createInvitation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
            examples:
              defaultExpiry:
                value: {}
              customExpiry:
                value:
                  expires_at: "2026-09-16T00:00:00Z"
      responses:
        "201":
          description: The signed invitation token and its metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Caller is not an org admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                notAdmin:
                  value: { message: "only org admins can create invitations" }
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/feedback-periods:
    post:
      tags: [Feedback Periods]
      summary: Create a feedback period
      description: |
        Opens a feedback period for the authenticated employee's organization.
        The caller must have `role: "org_admin"`; any other role returns `403`.
        The `organization_name` is taken from the caller's JWT (not the body),
        so a client cannot open a period for an org they do not belong to.
      operationId: createFeedbackPeriod
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedbackPeriodRequest"
            examples:
              h2:
                summary: H2 2026
                value:
                  name: H2 2026
                  start_date: "2026-07-01T00:00:00Z"
                  end_date: "2026-12-31T23:59:59Z"
      responses:
        "201":
          description: The created feedback period.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackPeriodResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Caller is not an org admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                notAdmin:
                  value: { message: "only org admins can create feedback periods" }
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Feedback Periods]
      summary: List feedback periods
      description: |
        Returns the feedback periods for the authenticated employee's
        organization, ordered by start date descending (most recent first).
        Any authenticated employee may list periods (an employee needs to see
        periods in order to file feedback against them). The organization is
        taken from the caller's JWT, so an employee can only see their own
        organization's periods.
      operationId: listFeedbackPeriods
      security:
        - BearerAuth: []
      responses:
        "200":
          description: The feedback periods for the caller's organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackPeriodListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/feedbacks:
    post:
      tags: [Feedback]
      summary: File a feedback entry
      description: |
        Files a feedback entry for a colleague. The reviewer is the
        authenticated employee (taken from the JWT subject), so `reviewer_id`
        in the body is ignored. A reviewer cannot review themselves.

        Validation:
          * `period_id` is required and must refer to an existing feedback
            period (the service validates it before persisting).
          * The period's date window must be open: now must fall within
            [start_date, end_date]. A not-yet-open or already-ended period
            returns `422`.
          * `reviewee_id` is required and must differ from the reviewer.
          * Each of the six score fields must be an integer in [1, 5].
          * `visibility` defaults to `"anonymous"` when omitted or empty, and
            must otherwise be one of `anonymous`, `named`.

        The created entry has `status: "submitted"`. To write feedback
        incrementally instead, start a draft via `POST /v1/feedback-drafts`.
      operationId: createFeedback
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedbackRequest"
            examples:
              sample:
                value:
                  period_id: 0190abcd-0000-7000-8000-000000000000
                  reviewee_id: 0190bbbb-0000-7000-8000-000000000000
                  communication_score: 4
                  leadership_score: 5
                  technical_score: 3
                  collaboration_score: 4
                  delivery_score: 5
                  trust_score: 2
                  strengths_comment: great teammate
                  weaknesses_comment: could document more
                  visibility: anonymous
      responses:
        "201":
          description: The created feedback entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: The period's date window is not open (not started or already ended).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                closed:
                  value: { message: "feedback period is not open for submission" }
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/me/feedbacks:
    get:
      tags: [Feedback]
      summary: List feedback received by the current employee
      description: |
        Returns one page of feedback entries received by the authenticated
        employee (i.e. entries others have written about them), ordered by
        `created_at` descending (newest first). The reviewee is the
        authenticated employee (taken from the JWT subject), so an employee can
        only list their own received feedback.

        Pagination is cursor-based: pass the `next_cursor` value from one
        response as the `cursor` query parameter on the next request. When
        `next_cursor` is `null` there are no more pages. An unknown cursor
        (one that does not refer to an existing feedback entry) returns `400`.

        Visibility policy: for entries with `visibility: "anonymous"`, the
        reviewer's identity is hidden from the reviewee (the caller of this
        endpoint) — `reviewer_id` is empty in the response. Entries with
        `visibility: "named"` include `reviewer_id` as usual.
      operationId: listMyFeedbacks
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListCursor"
      responses:
        "200":
          description: One page of received feedback plus the next-page cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackListResponse"
        "400":
          description: Unknown cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                unknownCursor:
                  value: { message: "unknown cursor" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/me/given-feedbacks:
    get:
      tags: [Feedback]
      summary: List feedback submitted by the current employee
      description: |
        Returns one page of feedback entries the authenticated employee has
        submitted (i.e. entries they wrote as the reviewer), ordered by
        `created_at` descending (newest first). The reviewer is the
        authenticated employee (taken from the JWT subject), so an employee
        can only list the feedback they gave themselves. Only submitted
        entries are listed; drafts stay in `GET /v1/feedback-drafts`.

        Pagination is cursor-based: pass the `next_cursor` value from one
        response as the `cursor` query parameter on the next request. When
        `next_cursor` is `null` there are no more pages. An unknown cursor
        (one that does not refer to an existing feedback entry) returns `400`.

        Visibility policy: none applies here — the caller is the reviewer of
        every entry, so `reviewer_id` is always included, even for anonymous
        entries (an author always knows their own identity).
      operationId: listMyGivenFeedbacks
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListCursor"
      responses:
        "200":
          description: One page of given feedback plus the next-page cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackGivenListResponse"
        "400":
          description: Unknown cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                unknownCursor:
                  value: { message: "unknown cursor" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/feedback-drafts/{id}/submit:
    post:
      tags: [Feedback Drafts]
      summary: Submit one of the caller's drafts
      description: |
        Transitions the caller's draft to the submitted state and releases its
        (reviewer, reviewee, period) slot. The request body is optional; when
        present it carries the same shape as the update request and its fields
        are applied before validation, letting a client submit and final-edit
        in one call.

        Full validation runs here:
          * All six score fields must be present and in [1, 5] (`400`).
          * `strengths_comment` and `weaknesses_comment` must be non-empty (`400`).
          * The period's date window must be open — now must fall within
            [start_date, end_date] (`422`).

        On success the entry behaves exactly like one created via
        `POST /v1/feedbacks`: it appears in the reviewee's and their manager's
        listings (subject to the visibility policy).
      operationId: submitFeedbackDraft
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DraftId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedbackDraftRequest"
            examples:
              finalEdit:
                summary: Fill in the last fields and submit
                value:
                  communication_score: 4
                  leadership_score: 5
                  technical_score: 3
                  collaboration_score: 4
                  delivery_score: 5
                  trust_score: 2
                  strengths_comment: great teammate
                  weaknesses_comment: could document more
                  visibility: named
      responses:
        "200":
          description: The submitted feedback entry, with status "submitted".
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No draft matches the ID for this caller (or it was already submitted).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                notFound:
                  value: { message: "feedback draft not found" }
        "409":
          description: The draft changed after this request read it; fetch it again and retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                concurrent:
                  value: { message: "draft was modified concurrently; fetch it again and retry" }
        "422":
          description: The period's date window is not open (not started or already ended).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                closed:
                  value: { message: "feedback period is not open for submission" }
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  responses:
...
    ValidationError:
      description: |
        Malformed request body or a failed validation. The `message` carries
        the specific reason (e.g. `"name is required"`,
        `"end_date must be after start_date"`,
        `"period_id does not refer to an existing feedback period"`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            invalidBody:
              value: { message: "invalid request body" }
            validation:
              value: { message: "name is required" }
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            internal:
              value: { message: "failed to register employee" }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A signed JWT obtained from `POST /v1/login`. Send as
        `Authorization: Bearer <access_token>`.
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: A human-readable description of the error.
      required: [message]
...
    RegisterEmployeeRequest:
      type: object
      description: Request body for `POST /v1/register`.
      properties:
        name:
          type: string
          description: Trimmed; must not be empty.
        organization_name:
          type: string
          description: |
            Required when `invite_token` is empty; ignored otherwise. Without
            a token it names a brand-new organization the caller bootstraps
            as its first admin — registering into an organization that
            already has members is rejected with 409.
        role:
          $ref: "#/components/schemas/Role"
          description: Omitted/ignored on input — the server assigns it from `invite_token`.
        manager_id:
          type: string
          format: uuid
          nullable: true
          description: Employee ID of the user's manager, or null.
        title:
          type: string
          description: Job title. Defaults to "".
        email:
          type: string
          format: email
          description: Stored lowercased; uniqueness is global across organizations.
        password:
          type: string
          format: password
          description: Plaintext; max 64 characters. Hashed server-side and never returned.
          maxLength: 64
        invite_token:
          type: string
          description: |
            If empty, the employee is created with `role: "org_admin"` (the
            first admin). If a valid invitation token is supplied, `role: "user"`.
          default: ""
      required: [name, organization_name, email, password]
...